updated code

+++ CMSC 127 Project Files/groups.py
@@ -26,8 +26,12 @@ def getGroups(userChoice):
 def viewAllGroups(userChoice):
     groups = getGroups(userChoice)
     
-    for key, value in groups.items():
-        print(key, "-", value)
+    if len(groups) > 0:
+        print("Your group/s: ")
+        for key, value in groups.items():
+            print(key, "-", value)
+    else:
+        print("You have no groups yet.")
 
 def createGroup(userChoice):
     print("\t-----ADDING-----\n")
@@ -53,15 +57,33 @@ def deleteGroup(userChoice):
     groups = getGroups(userChoice)
     viewAllGroups(userChoice)
 
-    print("\t-----DELETING-----\n")
-    groupToDelete = int(input("Enter group id: "))
-    if groupToDelete not in groups.keys():
-        print("The group is not existing!")
-    else:
-        query = f"DELETE FROM grp WHERE group_id = {groupToDelete}"
-        cursor.execute(query)
-        connection.commit()
-        print("Succecfully deleted group",groups[groupToDelete])
+    if len(groups) > 0:
+        print("\t-----DELETING-----\n")
+        groupToDelete = int(input("Enter group id: "))
+
+        if groupToDelete not in groups.keys():
+            print("The group is not existing!")
+        else:
+            # check if group has unsettled expenses
+            query = f"""select distinct (grp.group_id) from grp 
+            join group_has_expense h on grp.group_id = h.group_id 
+            join expense e on h.expense_id = e.expense_id 
+            where e. isSettled = 0 and grp.group_id = {groupToDelete}"""
+            cursor.execute(query)
+
+            result = cursor.fetchone()
+            if result is None:
+                # delete friends from belongs to
+                query = f"DELETE FROM belongsTo WHERE group_id = {groupToDelete}"
+                cursor.execute(query)
+                connection.commit()
+
+                query = f"DELETE FROM grp WHERE group_id = {groupToDelete}"
+                cursor.execute(query)
+                connection.commit()
+                print("Successfully deleted group",groups[groupToDelete])
+            else:
+                print("Fail to delete group. This group has unsettled expenses.")
 
 
 def searchGroup(userChoice):
@@ -90,32 +112,45 @@ def searchGroup(userChoice):
 
 def updateGroup(userChoice):
     groups = getGroups(userChoice)
-    print("\t-----UPDATING-----\n")
-    for key, value in groups.items():
-        print(key,"-",value)
-    group_id = int(input("Enter group id: "))
 
-    if group_id not in groups.keys():
-        print("The group is not existing!")
-    else:
-        print(f"\t-----UPDATING {groups[group_id]}-----\n")
+    if len(groups) > 0:
+        print("\t-----UPDATING-----\n")
+        for key, value in groups.items():
+            print(key,"-",value)
+        
         while True:
-            print(f"[1] Add a friend to a {groups[group_id]}")
-            print(f"[2] Remove a friend from {groups[group_id]}")
-            print(f"[3] Rename {groups[group_id]}")
-            print("[0] Back")
-            action = int(input("Enter the number of the action you wish to perform: "))
-
-            if action == 0:
+            try:
+                group_id = int(input("Enter group id: "))
                 break
-            elif action == 1:
-                addFriendToGroup(userChoice, group_id, groups[group_id])
-            elif action == 2:
-                removeFriendFromGroup()
-            elif action == 3:
-                renameGroup(group_id, groups[group_id])
-            else:
-                print("Invalid Choice!")
+            except ValueError:
+                print("Please enter integers only!")
+
+        if group_id not in groups.keys():
+            print("The group is not existing!")
+        else:
+            print(f"\t-----UPDATING {groups[group_id]}-----\n")
+            while True:
+                print(f"[1] Add a friend to a {groups[group_id]}")
+                print(f"[2] Remove a friend from {groups[group_id]}")
+                print(f"[3] Rename {groups[group_id]}")
+                print("[0] Back")
+                action = int(input("Enter the number of the action you wish to perform: "))
+
+                if action == 0:
+                    break
+                elif action == 1:
+                    addFriendToGroup(userChoice, group_id, groups[group_id])
+                    break
+                elif action == 2:
+                    removeFriendFromGroup(userChoice, group_id, groups[group_id])
+                    break
+                elif action == 3:
+                    renameGroup(group_id, groups[group_id])
+                    break
+                else:
+                    print("Invalid Choice!")
+    else:
+        print("No groups yet!")
     # Rename the group, add friend to the group, remove friend to a group
 
 def addFriendToGroup(userId, groupId, groupName):
@@ -124,13 +159,36 @@ def addFriendToGroup(userId, groupId, groupName):
     for key,value in friendsOfUser.items():
         print(key, "-", value)
 
-    friendId = int(input("Enter friend ID: "))
-    if friendId in friendsOfUser.keys():
-        query = f"INSERT INTO belongsTo (user_id,group_id) VALUES ({friendId},{groupId})"
-        cursor.execute(query)
-        print(f"{friendsOfUser[friendId]} has been added to {groupName}")
-    else:
-        print("Friend does not exist!")
+    while True:
+        try:
+            friendId = int(input("Enter friend ID (Enter 0 to exit): "))
+
+            if friendId == 0:
+                print("Exiting...")
+                break
+            elif friendId in friendsOfUser.keys():
+                # get all the members of the group
+                query = f"SELECT user_id FROM belongsTo where group_id = {groupId}"
+                cursor.execute(query)
+
+                membersOfTheGroup = []
+                result = cursor.fetchall()
+
+                for group in result:
+                    membersOfTheGroup.append(group[0])
+
+                if friendId not in membersOfTheGroup:
+                    query = f"INSERT INTO belongsTo (user_id,group_id) VALUES ({friendId},{groupId})"
+                    cursor.execute(query)
+                    connection.commit()
+                    print(f"{friendsOfUser[friendId]} has been added to {groupName}")
+                else:
+                    print("Friend is already a member of that group.")
+            else:
+                print("Friend does not exist!")
+        except ValueError:
+            print("Please enter integers only!")
+
 
 
 def removeFriendFromGroup(userId, groupId, groupName):
@@ -138,16 +196,26 @@ def removeFriendFromGroup(userId, groupId, groupName):
     friendsOfUser = friends.getFriends(userId)
     for key,value in friendsOfUser.items():
         print(key, "-", value)
-    friendId = int(input("Enter friend ID: "))
+    while True:
+        try:
+            friendId = int(input("Enter friend ID (Enter 0 to exit): "))
 
-    if friendId in friendsOfUser.keys():
-        query = f"DELETE FROM belongsTo (user_id,group_id) VALUES ({friendId},{groupId})"
-        cursor.execute(query)
-        print(f"{friendsOfUser[friendId]} has been removed from {groupName}")
-        # should catch the case where the friend cannot be removed from the group if the group still
-        #  has an unsettled expense
-    else:
-        print("Friend does not exist!")
+            if friendId == 0:
+                print("Exiting...")
+                break
+            elif friendId in friendsOfUser.keys():
+                query = f"DELETE FROM belongsTo where user_id = {friendId} and group_id = {groupId}"
+                cursor.execute(query)
+                connection.commit()
+                print(f"{friendsOfUser[friendId]} has been removed from {groupName}")
+                break
+                # should catch the case where the friend cannot be removed from the group if the group still
+                #  has an unsettled expense
+            else:
+                print("Friend does not exist!")
+            
+        except ValueError:
+            print("Enter integers only!")
 
 def renameGroup(groupId, oldGroupName):
     newGroupName = input("Enter new group name: ")
